+++ hazmat/init.go
@@ -211,7 +211,6 @@ func runStatus(full bool) error {
 		fmt.Println("  Next step: hazmat init")
 	}
 	fmt.Println()
-	maybeNotifyUpdateAvailable(os.Stderr)
 
 	if full {
 		ui := &UI{}